+++ Assets/Resources/CSV/CSVQuestionData.csv
@@ -1,10 +1,10 @@
-﻿question,answer1,answer2,answer3,answer4,correctAnswer
-left,left,down,right,up,left
-down,left,down,right,up,down
-left,left,down,right,up,left
-up,left,down,right,up,up
-right,left,down,right,up,right
-left,left,down,right,up,left
-down,left,down,right,up,down
-down,left,down,right,up,down
-left,left,down,right,up,left
+question,answer1,answer2,answer3,answer4,correctAnswer
+q1,correct,a1-2,a1-3,a1-4,correct
+q2,a2-1,correct,a2-3,a2-4,correct
+q3,a3-1,a3-2,correct,a3-4,correct
+q4,a4-1,a4-2,a4-3,correct,correct
+q5,correct,a5-2,a5-3,a5-4,correct
+q6,a6-1,correct,a6-3,a6-4,correct
+q7,a7-1,a7-2,correct,a7-4,correct
+q8,a8-1,a8-2,a8-3,correct,correct
+q9,correct,a9-2,a9-3,a9-4,correct
