+++ Assets/Settings/Scenes/Scenes/Demo.unity
@@ -1390,15 +1390,6 @@ MonoBehaviour:
   answerText2: {fileID: 334386987}
   answerText3: {fileID: 1059704933}
   answerText4: {fileID: 1780653712}
-  MaruBatsu1: {fileID: 1870022522}
-  MaruBatsu2: {fileID: 1763710367}
-  MaruBatsu3: {fileID: 94317982}
-  MaruBatsu4: {fileID: 200479364}
-  Next: {fileID: 1050571131}
-  answerBtn1: {fileID: 169619578}
-  answerBtn2: {fileID: 633890661}
-  answerBtn3: {fileID: 1833329379}
-  answerBtn4: {fileID: 843090455}
 --- !u!114 &841134403
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1683,7 +1674,7 @@ GameObject:
   - component: {fileID: 1050571136}
   - component: {fileID: 1050571135}
   - component: {fileID: 1050571134}
-  - component: {fileID: 1050571133}
+  - component: {fileID: 1050571137}
   m_Layer: 5
   m_Name: Next
   m_TagString: Untagged
@@ -1711,32 +1702,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 160, y: 30}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &1050571133
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1050571131}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 437c409653fee3d409abf41460352e48, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  questionText: {fileID: 143528543}
-  answerText1: {fileID: 1410531699}
-  answerText2: {fileID: 334386987}
-  answerText3: {fileID: 1059704933}
-  answerText4: {fileID: 1780653712}
-  MaruBatsu1: {fileID: 1870022522}
-  MaruBatsu2: {fileID: 1763710367}
-  MaruBatsu3: {fileID: 94317982}
-  MaruBatsu4: {fileID: 200479364}
-  Next: {fileID: 1050571131}
-  answerBtn1: {fileID: 169619578}
-  answerBtn2: {fileID: 633890661}
-  answerBtn3: {fileID: 1833329379}
-  answerBtn4: {fileID: 843090455}
 --- !u!114 &1050571134
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1781,8 +1746,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1050571133}
-        m_TargetAssemblyTypeName: GameManager, Assembly-CSharp
+      - m_Target: {fileID: 1050571137}
+        m_TargetAssemblyTypeName: QuestionManager, Assembly-CSharp
         m_MethodName: NextQuestion
         m_Mode: 1
         m_Arguments:
@@ -1831,6 +1796,32 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1050571131}
   m_CullTransparentMesh: 1
+--- !u!114 &1050571137
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1050571131}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d60c731691213824ba823d044f8bbc5d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  questionText: {fileID: 143528543}
+  answerText1: {fileID: 1410531699}
+  answerText2: {fileID: 334386987}
+  answerText3: {fileID: 1059704933}
+  answerText4: {fileID: 1780653712}
+  answerBtn1: {fileID: 169619578}
+  answerBtn2: {fileID: 633890661}
+  answerBtn3: {fileID: 1833329379}
+  answerBtn4: {fileID: 843090455}
+  MaruBatsu1: {fileID: 1870022522}
+  MaruBatsu2: {fileID: 1763710367}
+  MaruBatsu3: {fileID: 94317982}
+  MaruBatsu4: {fileID: 200479364}
+  Next: {fileID: 1050571131}
 --- !u!1 &1059704931
 GameObject:
   m_ObjectHideFlags: 0
