+++ Assets/Settings/Scenes/Scripts/QuestionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d60c731691213824ba823d044f8bbc5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
